requirements.txt
# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0

# MEG processing
mne>=1.0.0
mne-bids>=0.10

# Visualization
matplotlib>=3.5.0

# Spatial statistics (optional, for spin tests)
brainspace>=0.1.0

# Configuration
pyyaml>=6.0

# Development (optional)
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
